requirements.txt

Flask==3.1.3
Flask-Login==0.6.2
motor==3.4.0
pymongo==4.6.0
python-dotenv==1.0.0
certifi==2024.1.1
werkzeug==3.0.0
flask-mail==0.9.1
PyJWT==2.8.0
google-generativeai==1.2.1
